# m2m-wallet/statics/swagger/api.swagger.json
{"swagger":"2.0","basePath":"","info":{"title":"M2M wallet REST API","version":"1.0.0","description":""},"schemes":null,"consumes":["application/json"],"produces":["application/json"],"paths":{"/api/money/active_account/{money_abbr}":{"get":{"operationId":"GetActiveMoneyAccount","parameters":[{"enum":["Ether"],"in":"path","name":"money_abbr","required":true,"type":"string"},{"format":"int64","in":"query","name":"orgId","required":false,"type":"string"}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/apiGetActiveMoneyAccountResponse"}}},"tags":["MoneyService"]}},"/api/money/change_history/{money_abbr}":{"get":{"operationId":"GetChangeMoneyAccountHistory","parameters":[{"enum":["Ether"],"in":"path","name":"money_abbr","required":true,"type":"string"},{"format":"int64","in":"query","name":"orgId","required":false,"type":"string"},{"format":"int64","in":"query","name":"offset","required":false,"type":"string"},{"format":"int64","in":"query","name":"limit","required":false,"type":"string"}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/apiGetMoneyAccountChangeHistoryResponse"}}},"tags":["MoneyService"]}},"/api/money/modify_account/{money_abbr}":{"put":{"operationId":"ModifyMoneyAccount","parameters":[{"enum":["Ether"],"in":"path","name":"money_abbr","required":true,"type":"string"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/apiModifyMoneyAccountRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/apiModifyMoneyAccountResponse"}}},"tags":["MoneyService"]}},"/api/supernode/active_account/{money_abbr}":{"get":{"operationId":"GetSuperNodeActiveMoneyAccount","parameters":[{"enum":["Ether"],"in":"path","name":"money_abbr","required":true,"type":"string"}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/apiGetSuperNodeActiveMoneyAccountResponse"}}},"tags":["SuperNodeService"]}},"/api/top_up/history":{"get":{"operationId":"GetTopUpHistory","parameters":[{"format":"int64","in":"query","name":"orgId","required":false,"type":"string"},{"format":"int64","in":"query","name":"offset","required":false,"type":"string"},{"format":"int64","in":"query","name":"limit","required":false,"type":"string"}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/apiGetTopUpHistoryResponse"}}},"tags":["TopUpService"]}},"/api/wallet/balance/{org_id}":{"get":{"operationId":"GetWalletBalance","parameters":[{"format":"int64","in":"path","name":"org_id","required":true,"type":"string"}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/apiGetWalletBalanceResponse"}}},"tags":["WalletService"]}},"/api/wallet/vmxc/tx_history":{"get":{"operationId":"GetVmxcTxHistory","parameters":[{"format":"int64","in":"query","name":"orgId","required":false,"type":"string"},{"format":"int64","in":"query","name":"offset","required":false,"type":"string"},{"format":"int64","in":"query","name":"limit","required":false,"type":"string"}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/apiGetVmxcTxHistoryResponse"}}},"tags":["WalletService"]}},"/api/withdraw/history":{"get":{"operationId":"GetWithdrawHistory","parameters":[{"in":"query","name":"orgId","required":false,"type":"string"},{"in":"query","name":"offset","required":false,"type":"string"},{"format":"int64","in":"query","name":"limit","required":false,"type":"string"}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/apiGetWithdrawHistoryResponse"}}},"tags":["WithdrawService"]}},"/api/withdraw/{money_abbr}":{"get":{"operationId":"GetWithdrawFee","parameters":[{"description":"type of crypto currency","enum":["Ether"],"in":"path","name":"money_abbr","required":true,"type":"string"}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/apiGetWithdrawFeeResponse"}}},"summary":"Get data for current withdraw fee","tags":["WithdrawService"]},"post":{"operationId":"WithdrawReq","parameters":[{"enum":["Ether"],"in":"path","name":"money_abbr","required":true,"type":"string"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/apiWithdrawReqRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/apiWithdrawReqResponse"}}},"tags":["WithdrawService"]}}},"definitions":{"apiGetActiveMoneyAccountResponse":{"properties":{"activeAccount":{"type":"string"},"error":{"type":"string"}},"type":"object"},"apiGetMoneyAccountChangeHistoryResponse":{"properties":{"changeHistory":{"items":{"$ref":"#/definitions/apiMoneyAccountChangeHistory"},"type":"array"},"count":{"format":"int64","type":"string"},"error":{"type":"string"}},"type":"object"},"apiGetSuperNodeActiveMoneyAccountResponse":{"properties":{"Error":{"type":"string"},"supernodeActiveAccount":{"type":"string"}},"type":"object"},"apiGetTopUpHistoryResponse":{"properties":{"count":{"format":"int64","type":"string"},"error":{"type":"string"},"topupHistory":{"items":{"$ref":"#/definitions/apiTopUpHistory"},"type":"array"}},"type":"object"},"apiGetVmxcTxHistoryResponse":{"properties":{"count":{"format":"int64","type":"string"},"error":{"type":"string"},"txHistory":{"items":{"$ref":"#/definitions/apiVmxcTxHistory"},"type":"array"}},"type":"object"},"apiGetWalletBalanceResponse":{"properties":{"balance":{"format":"float","type":"number"},"error":{"type":"string"}},"type":"object"},"apiGetWithdrawFeeResponse":{"properties":{"error":{"type":"string"},"withdrawFee":{"description":"Withdraw object.","format":"float","type":"number"}},"type":"object"},"apiGetWithdrawHistoryResponse":{"properties":{"count":{"format":"int64","type":"string"},"error":{"type":"string"},"withdrawHistory":{"items":{"$ref":"#/definitions/apiWithdrawHistory"},"type":"array"}},"type":"object"},"apiModifyMoneyAccountRequest":{"properties":{"currentAccount":{"type":"string"},"moneyAbbr":{"$ref":"#/definitions/apiMoney"},"orgId":{"format":"int64","type":"string"}},"type":"object"},"apiModifyMoneyAccountResponse":{"properties":{"error":{"type":"string"},"status":{"format":"boolean","type":"boolean"}},"type":"object"},"apiMoney":{"default":"Ether","enum":["Ether"],"type":"string"},"apiMoneyAccountChangeHistory":{"properties":{"createdAt":{"type":"string"},"from":{"type":"string"},"to":{"type":"string"}},"type":"object"},"apiTopUpHistory":{"properties":{"amount":{"format":"float","type":"number"},"createdAt":{"type":"string"},"from":{"type":"string"},"to":{"type":"string"}},"type":"object"},"apiVmxcTxHistory":{"properties":{"amount":{"format":"float","type":"number"},"createdAt":{"type":"string"},"from":{"type":"string"},"to":{"type":"string"},"txType":{"type":"string"}},"type":"object"},"apiWithdrawHistory":{"properties":{"amount":{"format":"float","type":"number"},"createdAt":{"type":"string"},"from":{"type":"string"},"money":{"$ref":"#/definitions/apiMoney"},"to":{"type":"string"}},"type":"object"},"apiWithdrawReqRequest":{"properties":{"amount":{"format":"float","type":"number"},"moneyAbbr":{"$ref":"#/definitions/apiMoney"},"orgId":{"format":"int64","type":"string"}},"type":"object"},"apiWithdrawReqResponse":{"properties":{"error":{"type":"string"},"status":{"format":"boolean","type":"boolean"}},"type":"object"}}}